% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_apchagi.R
\docType{data}
\name{all_apchagi}
\alias{all_apchagi}
\title{Front kick for Gaetan, Bleuenn and Maxime}
\format{
A data frame with 3696 rows and 8 variables:
\describe{
\item{time}{The time recorded in second}
\item{loc}{The localisation, i.e. the joint}
\item{x}{The coordinate on the x-axis}
\item{y}{The coordinate on the y-axis}
\item{z}{The coordinate on the z-axis}
\item{frame}{The index of the frame}
\item{fps}{The number of frames per second}
\item{name}{The name of the practitioner}
}
}
\usage{
all_apchagi
}
\description{
In this data set, we have recorded three practitioners while they were doing a front kick.
}
\keyword{datasets}
