% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaetan_apchagi.R
\docType{data}
\name{gaetan_apchagi}
\alias{gaetan_apchagi}
\title{Front kick for Gaetan}
\format{
A data frame with 1232 rows and 8 variables:
\describe{
\item{time}{The time}
\item{loc}{The localisation, i.e. the joint}
\item{x}{The coordinate on the x-axis}
\item{y}{The coordinate on the y-axis}
\item{z}{The coordinate on the z-axis}
\item{frame}{The index of the frame}
\item{fps}{The number of frame per second}
\item{name}{The name of the figure}
}
}
\usage{
gaetan_apchagi
}
\description{
In this data set, we have recorded Gaetan while he was doing a front kick.
}
\keyword{datasets}
