% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaetan_apchagi_mediapipe.R
\docType{data}
\name{gaetan_apchagi_mediapipe}
\alias{gaetan_apchagi_mediapipe}
\title{Mediapipe data set}
\format{
A data frame with 6831 rows and 4 variables:
\describe{
\item{time}{The time}
\item{loc}{The localisation, i.e. the joint}
\item{coord}{The dimension on which the value is measured, x, y and z}
\item{value}{The corrdinate in a 3 dimensional space}
}
}
\usage{
gaetan_apchagi_mediapipe
}
\description{
This data set is the one obtained from MediaPipe after analyzing a video
}
\keyword{datasets}
