% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human.R
\docType{data}
\name{human}
\alias{human}
\title{Weekly electricity consumption in France from 1996 to 2009 in MW
meteo and socio-economic variables related to it}
\format{
A data frame with 731 rows and 11 variables:
\describe{
\item{segment}{time index, in number of weeks}
\item{extr1}{Day}
\item{extr2}{Month}
}
}
\usage{
human
}
\description{
Weekly electricity consumption in France from 1996 to 2009 in MW
meteo and socio-economic variables related to it
}
\keyword{datasets}
