% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidekick.R
\name{sidekick}
\alias{sidekick}
\title{Create supplementary data to illustrate the representation of the structure}
\usage{
sidekick(
  joint,
  num.joint,
  num.frame = NULL,
  num.name,
  num.x,
  num.y,
  joint1,
  joint2 = NULL
)
}
\arguments{
\item{joint}{The joint dataset: the coordinates of the joints as a function of time}

\item{num.joint}{The index of the column associated with the joint variable}

\item{num.frame}{The index of the column associated with the frame variable}

\item{num.name}{The index of the column associated with the name variable}

\item{num.x}{The index of the column associated with the x-axis variable represented on the graphical output}

\item{num.y}{The index of the column associated with the y-axis variable represented on the graphical output}

\item{joint1}{A singleton, a couple or a triplet of joints}

\item{joint2}{A singleton, a couple or a triplet of joints}
}
\value{
A dataset formatted to be plotted with the superskeleton function
}
\description{
This function creates a new dataset for the superskeleton function.
}
\examples{
data(gaetan_apchagi)
# To represent the ankle measured on the y-axis as a function of time
S1_right_ankle <- sidekick(joint=gaetan_apchagi, num.joint=2, num.name=8,
num.x=6, num.y=4, joint1="RIGHT_ANKLE", joint2=NULL)

# To represent the difference between the knee and the ankle
# measured on the y-axis as a function of time
C1_right_knee_ankle <- sidekick(joint=gaetan_apchagi, num.joint=2, num.name=8,
num.x=6, num.y=4, joint1=c("RIGHT_KNEE", "RIGHT_ANKLE"),
joint2=NULL)

# To represent the angle of the knee measured on the plane defined
# by the x-axis and the y-axis
T1_right_knee_ankle <- sidekick(joint=gaetan_apchagi, num.joint=2, num.name=8,
num.x=6, num.y=c(3,4), joint1=c("RIGHT_HIP","RIGHT_KNEE", "RIGHT_ANKLE"),
joint2=NULL)

# To represent the trajectory of the ankle with respect to the knee
# measured both on the y-axis as a function of the time
S2_right_ankle_knee <- sidekick(joint=gaetan_apchagi, num.joint=2, num.name=8,
num.x=4, num.y=4, joint1="RIGHT_ANKLE", joint2="RIGHT_KNEE")

}
