% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleton.R
\name{skeleton}
\alias{skeleton}
\title{Title}
\usage{
skeleton(
  joint,
  structure,
  num.joint,
  num.frame,
  num.x,
  num.y,
  num.name,
  frame.index = NULL,
  body.part = "",
  color.part = "black",
  fps = 30
)
}
\arguments{
\item{joint}{The joint dataset: the coordinates of the joints as a function of time}

\item{structure}{The structure dataset: a first column with the segments composing the structure, two other columns defining the extremities of the segments}

\item{num.joint}{The index of the column associated with the joint variable}

\item{num.frame}{The index of the column associated with the frame variable}

\item{num.x}{The index of the column associated with the x-axis variable represented on the graphical output}

\item{num.y}{The index of the column associated with the y-axis variable represented on the graphical output}

\item{num.name}{The index of the column associated with the name variable}

\item{frame.index}{The index of the frame you want to represent (static representation)}

\item{body.part}{The names of the segments you want to represent}

\item{color.part}{The colour you want to use to represent the segments}

\item{fps}{The number of frames per second}
}
\value{
An animation by default or a static representation for a given frame
}
\description{
Title
}
\examples{

\dontrun{
data(gaetan_apchagi)
data(human)
skeleton(joint=gaetan_apchagi, structure= human, num.joint=2,
num.frame=6, num.x=3, num.y=4, num.name=8)

skeleton(joint=gaetan_apchagi, structure= human, num.joint=2,
num.frame=6, num.x=3, num.y=4, num.name=8, body.part = "tibia_r",color.part = "orange")
}

data(gaetan_apchagi)
data(human)

skeleton(joint=gaetan_apchagi, structure= human, num.joint=2,
num.frame=6, num.x=3, num.y=4, num.name=8, body.part = "tibia_r",
frame.index=25, color.part = "orange")

}
