% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beyondkick.R
\name{beyondkick}
\alias{beyondkick}
\title{Title}
\usage{
beyondkick(
  joint,
  num.joint,
  num.frame = NULL,
  num.x,
  num.y,
  origin,
  seg1,
  seg2 = NULL
)
}
\arguments{
\item{joint}{The joint dataset: the coordinates of the joints as a function of time}

\item{num.joint}{The index of the column associated with the joint variable}

\item{num.frame}{The index of the column associated with the frame variable}

\item{num.x}{The index of the column associated with the x-axis variable represented on the graphical output}

\item{num.y}{The index of the column associated with the y-axis variable represented on the graphical output}

\item{origin}{The origin of the basis, a joint}

\item{seg1}{A triplet, the definition of the segment in terms of joints, and the coordinate on this segment}

\item{seg2}{A triplet, the definition of the segment in terms of joints, and the coordinate on this segment}
}
\value{
A supplementary dataset to be used with the beyondskeleton function
}
\description{
Title
}
\examples{
\dontrun{
data(gaetan_apchagi)
plexus <- beyondkick(joint=gaetan_apchagi, num.joint=2, num.x = 3, num.y =4,
num.frame=6, origin="RIGHT_SHOULDER", seg1 = c("RIGHT_SHOULDER", "LEFT_SHOULDER", 1/2),
seg2=c("RIGHT_SHOULDER", "RIGHT_HIP", 1/2))
}
}
