% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaetan_apchagi_mediapipe.R
\docType{data}
\name{gaetan_apchagi_mediapipe}
\alias{gaetan_apchagi_mediapipe}
\title{Weekly electricity consumption in France from 1996 to 2009 in MW
meteo and socio-economic variables related to it}
\format{
A data frame with 731 rows and 11 variables:
\describe{
\item{time}{time index, in number of weeks}
\item{loc}{Day}
\item{coord}{Month}
\item{value}{Year}
}
}
\usage{
gaetan_apchagi_mediapipe
}
\description{
Weekly electricity consumption in France from 1996 to 2009 in MW
meteo and socio-economic variables related to it
}
\keyword{datasets}
