% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human.R
\docType{data}
\name{human}
\alias{human}
\title{The description of the structure}
\format{
A data frame with 18 rows and 3 variables:
\describe{
\item{segment}{The name of the segment}
\item{extr1}{One of the extremities}
\item{extr2}{The other extremity}
}
}
\usage{
human
}
\description{
This data set is very important, as the structure of the "object" of interest is described
}
\keyword{datasets}
